another warning fixed.
ioeye-staging-commit :: 45f2f518f5d12f31c1302165e3e7b38e8537449b

+++ common/instamsg/driver/sg_stdlib.c
@@ -107,7 +107,7 @@ char* sg_strnstr(const char *str1, const char *str2, int maxSize)
     {
         return NULL;
     }
-    if(maxSize < strlen(str2))
+    if(maxSize < ((int)(strlen(str2))))
     {
         return NULL;
     }
